[monodroid] Refactor: AndroidSystem placed in separate files
The `AndroidSystem` class now lives in separate source files. No functional
changes made.
